refactor: 렌더링 관련 함수 분리
